move enums and types in dpp folder
