+++ pkg/src/wasm.ts
@@ -3,7 +3,7 @@ import { setDpp } from './dpp/dpp.js'
 
 setDpp(protocol)
 
-export { KeyType, Purpose, PlatformVersionWASM, NetworkWASM, SecurityLevel } from './enums.js'
+export { KeyType, Purpose, PlatformVersionWASM, NetworkWASM, SecurityLevel } from './dpp/enums.js'
 export { IdentityWASM, IdentifierWASM, IdentityPublicKeyWASM } from './dpp/dpp.js'
 
-export * from './types.js'
+export * from './dpp/types.js'
